move seq_derive and fix include paths, remove antimirov code
