+++ src/ast/rewriter/seq_derive.h
